DEV-2169 Map batch_id
There is a metadata field `batch_id` which is used by the batch intake
process to group a batch of items. Although this is just a logical grouping
identifier it is an important one as it is used for reporting and such.

We map the `batch_id` on the bag level and not the SIP itself. We map the ID
to a "meemoo internal identifier" in the `bag-info.txt`.