DEV-2184 Validate sidecar XML
Add a mechanism to check if the necessary fields in the sidecar are available.
If not, log an error, send a failed event to Pulsar and stop the processing.
For now, only the md5 is deemed necessary.

Also, if the local_id is not available in the sidecar, do not add it as
key/value in the data of the event.