DEV-2215 Add Meemoo-Workflow in bag-info
There is a very specific case in which the value of the `sp_name` needs
to be mapped to the mhs metadata. However, this is an exception, so the
mapping is not on the SIP level but on the bag level.

We map the `sp_name` value to the value of the `Meemoo-Workflow` key in
the `bag-info.txt`.