DEV-2291 Fix filename of zipped bag
The filename of the zipped bag should be derived from the filename of
the essence/sidecar.

For example, if the essence has the filename `file.111.mkv`, then the
filename of the zip should be `file.111.bag.zip`.

The `with_suffix` method in the Pathlib library however replaces the suffix
resulting in `file.bag.zip`.

Just append the filename with `.bag.zip`.